Refactor : deplacement de tous les composant comun dans comon
