+++ src/app/common/components/layout/footer/footer.component.ts
