+++ src/app/common/components/layout/footer/footers.component.theme.scss
@@ -1,5 +1,5 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
+@use '../../../../../../node_modules/@angular/material/index' as mat;
 
 @mixin color($theme) {
   // Get the color config from the theme.
