+++ src/app/common/components/layout/navbar/navbar.component.html
