+++ src/app/common/components/layout/navbar/navbar.desktop.component.scss
