+++ src/app/common/components/layout/navbar/navbar.mobile.component.scss
