+++ src/app/common/components/layout/navbar/theme-chooser/theme-chooser.component.html
